Update Provider Rating Count to int

+++ src/SFA.DAS.ApprenticeFeedback.Domain/Entities/FeedbackForProvidersResult.cs
@@ -7,7 +7,7 @@ namespace SFA.DAS.ApprenticeFeedback.Domain.Entities
     {
         public long Ukprn { get; set; }
         public string ProviderRating { get; set; }
-        public long ProviderRatingCount { get; set; }
+        public int ProviderRatingCount { get; set; }
         public string AttributeName { get; set; }
         public string Category { get; set; }
         public int ProviderAttributeAgree { get; set; }
